tests: cleanup temporary files
Call the configuration directory cleanup function before exiting
so there are no files left behind, even in case of error.

Add KeyFileExt to the list of files that are removed.

Reviewed-on: http://gerrit.openafs.org/8142
Tested-by: BuildBot <buildbot@rampaginggeek.com>

+++ tests/common/config.c
@@ -120,13 +120,16 @@ fail:
 void
 afstest_UnlinkTestConfig(char *dir)
 {
-    unlinkConfigFile(dir, "KeyFile");
-    unlinkConfigFile(dir, "CellServDB");
-    unlinkConfigFile(dir, "ThisCell");
-    unlinkConfigFile(dir, "UserList");
-    unlinkConfigFile(dir, "krb.conf");
-    unlinkConfigFile(dir, "krb.excl");
-    rmdir(dir);
+    if (getenv("MAKECHECK") != NULL) {
+	unlinkConfigFile(dir, "KeyFile");
+	unlinkConfigFile(dir, "KeyFileExt");
+	unlinkConfigFile(dir, "CellServDB");
+	unlinkConfigFile(dir, "ThisCell");
+	unlinkConfigFile(dir, "UserList");
+	unlinkConfigFile(dir, "krb.conf");
+	unlinkConfigFile(dir, "krb.excl");
+	rmdir(dir);
+    }
 }
 
 int
